LPS-105380 Move classes implementing 'ManagedServiceFactory' to package '*.configuration.admin.service'
